% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{capture.env}
\alias{capture.env}
\title{return output of \code{ls.str}}
\usage{
capture.env(env)
}
\arguments{
\item{env}{an environment}
}
\value{
the output of \code{ls.str(x)}, as a string
}
\description{
calls \code{print} and returns output as a string.  this is a utility 
function for use with watches in the js client.
}
\seealso{
\code{\link{capture.print}}, \code{\link{capture.str}}, \code{\link{capture.histogram}}
}

