% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{capture.histogram}
\alias{capture.histogram}
\title{return output of \code{histogram}}
\usage{
capture.histogram(x)
}
\arguments{
\item{x}{any R object}
}
\value{
the output of \code{hist(x, plot=F)}, as a string
}
\description{
calls \code{histogram} and returns output as a string, wihtout plotting
the histogram.  this is a utility function for use with watches in the 
js client.
}
\seealso{
\code{\link{capture.print}}, \code{\link{capture.str}}
}

