% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{capture.print}
\alias{capture.print}
\title{return output of \code{print}}
\usage{
capture.print(x)
}
\arguments{
\item{x}{any R object}
}
\value{
the output of \code{print(x)}, as a string
}
\description{
calls \code{print} and returns output as a string.  this is a utility 
function for use with watches in the js client.
}
\seealso{
\code{\link{capture.str}}, \code{\link{capture.histogram}}
}

