% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{capture.str}
\alias{capture.str}
\title{return output of \code{str}}
\usage{
capture.str(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
the output of \code{str(x)}, as a string
}
\description{
calls \code{str} and returns output as a string.  this is a utility 
function for use with watches in the js client.
}
\seealso{
\code{\link{capture.print}}, \code{\link{capture.histogram}}
}

