% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{close.js.client.progress.bar}
\alias{close.js.client.progress.bar}
\title{Close a progress bar handle}
\usage{
\method{close}{js.client.progress.bar}(pb)
}
\arguments{
\item{pb}{An object of class \code{js.client.progress.bar}}
}
\description{
implementation of generic \code{close} for objects of class 
\code{js.client.progress.bar}
}
\seealso{
\code{\link{js.client.progress.bar}}, \code{\link{js.client.get.progress.bar}},
\code{\link{js.client.set.progress.bar}}
}

