% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{device}
\alias{device}
\title{create a new graphics device}
\usage{
device(name = "json-ipc", background = "white", width = 600,
  height = 400, pointsize = 12)
}
\description{
create a new graphics device
}

