% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{device.resize}
\alias{device.resize}
\title{resize a graphics device}
\usage{
device.resize(device, width, height, replay = T)
}
\arguments{
\item{device}{A device ID}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{replay}{If True, replay the graphics device history (redraw)}
}
\description{
resize a graphics device
}

