% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{file.changed}
\alias{file.changed}
\title{Notification when a watched file has changed.}
\usage{
file.changed(path)
}
\arguments{
\item{path}{Path to the file}
}
\description{
Notification when a watched file has changed.
}

