% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{js.client.get.progress.bar}
\alias{js.client.get.progress.bar}
\title{Get progress bar value}
\usage{
js.client.get.progress.bar(pb)
}
\arguments{
\item{pb}{An object of class \code{js.client.progress.bar}}
}
\description{
Get progress bar value
}
\seealso{
\code{\link{js.client.progress.bar}}, \code{\link{js.client.set.progress.bar}}
}

