% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{js.client.progress.bar}
\alias{js.client.progress.bar}
\title{Create progress bar}
\usage{
js.client.progress.bar(min = 0, max = 1, initial = min, ...)
}
\arguments{
\item{min}{Minimum value}

\item{max}{Maximum value}

\item{initial}{Initial value.  Defaults to \code{min}}

\item{...}{Additional arguments (width, style)}
}
\value{
An object of class \code{js.client.progress.bar}
}
\description{
Create a progress bar and return the handle.
}
\seealso{
\code{\link{js.client.get.progress.bar}}, \code{\link{js.client.set.progress.bar}}
}

