% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{js.client.remove.watch}
\alias{js.client.remove.watch}
\title{Remove a watch variable}
\usage{
js.client.remove.watch(index)
}
\arguments{
\item{index}{the (integer) index into the list of watches}
}
\description{
Remove watch, by index.  watches are somewhat complicated because you
can have multiple watches on one field, and we allow a full set of equal
parameters; hence there's no reasonable way to look up a value.
}

