% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{js.client.set.progress.bar}
\alias{js.client.set.progress.bar}
\title{Set value in a progress bar}
\usage{
js.client.set.progress.bar(pb, value)
}
\arguments{
\item{pb}{An object of class \code{js.client.progress.bar}}

\item{value}{A number}
}
\description{
Set value in a progress bar
}
\seealso{
\code{\link{js.client.progress.bar}}, \code{\link{js.client.get.progress.bar}}
}

