% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{js.client.watch.file}
\alias{js.client.watch.file}
\title{Watch a file}
\usage{
js.client.watch.file(path, FUNC = NULL, override = T)
}
\arguments{
\item{path}{Path to the file}

\item{FUNC}{Code to execute when the file changes.  Defaults to \code{source(path)}.}

\item{override}{Remove any prior file watches.  You might add two separate watches
that execute different functions when a file changes; this is supported, unless the 
\code{override} parameter is set. Defaults to True.}
}
\description{
Watch a file and reload (or execute some arbitrary code) when the file changes.
}

