% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{locals}
\alias{locals}
\title{get locals}
\usage{
locals(envir)
}
\arguments{
\item{envir}{The current environment -- generally \code{.GlobalEnv}, but 
when debugging, the active function.}
}
\description{
the idea is to give a relatively short representation of each field, as
a string.  the first line is used as the brief description and the full
text is displayed for a selected item.
}
\details{
by default we're using \code{print} for functions and \code{str} for all other 
fields, and we have wrappers for these that call \code{capture.output}.
}

