% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsclientlib.R
\name{watches}
\alias{watches}
\title{Get watches}
\usage{
watches()
}
\value{
List of watche variables
}
\description{
Get watches.  A numbered list is returned, use those indexes to 
manipulate watches.
}
\seealso{
\code{\link{js.client.remove.watch}}
}

